Fix code location
